pyproject.toml
[tool.poetry]
name = "well-read"
version = "0.1.0"
description = ""
authors = ["Rounak Sen <rounaksen225@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"